+++ R/episodes.R
@@ -16,7 +16,10 @@ st_script_info <- function(){
   .season_sub <- function(x){
     name <- c("One", "Two", "Three", "Four", "Five", "Six", "Seven")
     suppressWarnings(
-      as.integer(paste0(sapply(1:7, function(i) if(grepl(name[i], x)) gsub(paste0(".*", name[i]), i, x) else ""), collapse = ""))
+      as.integer(paste0(sapply(1:7, function(i){
+        if(grepl(name[i], x)) gsub(paste0(".*", name[i]), i, x) else ""
+      }
+      ), collapse = ""))
     )
   }
   season <- ifelse(grepl("^Season .*", txt), sapply(txt, .season_sub), NA)
@@ -55,4 +58,47 @@ st_script_info <- function(){
     dplyr::arrange(.data[["format"]], .data[["series"]], .data[["number"]]) %>%
     dplyr::mutate(series = as.character(.data[["series"]])) %>%
     tidyr::fill(.data[["season"]])
+}
+
+#' Download episode scripts
+#'
+#' Download episode scripts into a nested tidy data frame.
+#'
+#' @param download_dir download directory.
+#' @param keep logical, if \code{FALSE} (default) then downloaded files are removed after processing.
+#' @param overwrite logical, if \code{FALSE} (default) then no downloading occurs for any file already present in \code{download_dir}.
+#'
+#' @return a data frame; may write files to disk.
+#' @export
+#'
+#' @examples
+#' \dontrun{x <- st_script_download("data-raw/episode_scripts", TRUE)}
+st_script_download <- function(download_dir = tempdir(), keep = FALSE, overwrite = FALSE){
+  x <- st_script_info() %>%
+    dplyr::mutate(
+      file = ifelse(.data[["format"]] == "episode",
+                    paste0(.data[["series"]], "-s", .data[["season"]], "e", .data[["number"]], "-",
+                           gsub(":|\\?", "", gsub(" ", "_", .data[["title"]])), ".txt"),
+                    paste0(.data[["series"]], "-mov", .data[["number"]], "-",
+                           gsub(":|\\?", "", gsub(" ", "_", .data[["title"]])), ".txt"))
+    )
+
+  x <- dplyr::filter(x, !is.na(.data[["url"]])) # temporary: need to update st_script_info missing entries
+
+  get_script <- function(i){
+    file <- file.path(download_dir, x$file[i])
+    if(!file.exists(file) | overwrite){
+      cat("Downloading ", basename(file), "...\n", sep = "")
+      z <- tryCatch(utils::download.file(x$url[i], file, quite = TRUE), error = function(e) "fail")
+      if(z == "fail") return(NA)
+    }
+
+    # TODO: process script
+    cat("Reading file: ", file, "...\n", sep = "")
+    txt <- readLines(file, warn = FALSE) # temporary, just returning the lines of raw text for now
+
+    if(!keep) unlink(file, recursive = TRUE, force = TRUE)
+    txt
+  }
+  dplyr::mutate(x, file = NULL, text = purrr::map(1:nrow(x), get_script))
 }