+++ tests/testthat/test-substitutions.R
@@ -1,0 +1,3 @@
+context("substitutions")
+
+expect_equal(romans_sub("a VI b"), "a 6 b")